pyproject.toml
[tool.poetry]
name = "plotbot"
version = "0.1.0"
description = "Collection of plotter drawings"
authors = ["Dan McKinley <mcfunley@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
numpy = "^1.22.3"
click = "^8.1.2"
scipy = "^1.8.0"
pycairo = "^1.21.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"

[tool.poetry.scripts]
plotbot = "plotbot.cli:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
